Move `proof.ts` to `lib/bitcoin`
This content of the `proof.ts` file actually belongs to `lib/bitcoin/proof.ts`.
Part of this code related to blocks, should live in `lib/bitcoin/block.ts`.
