+++ typescript/src/index.ts
@@ -1,5 +1,7 @@
 // TODO: Consider exports refactoring as per discussion https://github.com/keep-network/tbtc-v2/pull/460#discussion_r1084530007
 
+import { validateTransactionProof } from "./lib/bitcoin"
+
 import {
   calculateDepositAddress,
   getRevealedDeposit,
@@ -22,8 +24,6 @@ import {
   finalizeOptimisticMint,
   getOptimisticMintingRequest,
 } from "./optimistic-minting"
-
-import { validateTransactionProof } from "./proof"
 
 export const TBTC = {
   calculateDepositAddress,
