+++ typescript/src/lib/bitcoin/block.ts
@@ -1,5 +1,6 @@
 import { BigNumber } from "ethers"
 import { Hex } from "../utils"
+import { computeHash256, hashLEToBigNumber } from "./hash"
 
 /**
  * BlockHeader represents the header of a Bitcoin block. For reference, see:
@@ -140,4 +141,109 @@ export function targetToDifficulty(target: BigNumber): BigNumber {
     "0xffff0000000000000000000000000000000000000000000000000000"
   )
   return DIFF1_TARGET.div(target)
+}
+
+/**
+ * Validates a chain of consecutive block headers by checking each header's
+ * difficulty, hash, and continuity with the previous header. This function can
+ * be used to validate a series of Bitcoin block headers for their validity.
+ * @param blockHeaders An array of block headers that form the chain to be
+ *        validated.
+ * @param previousEpochDifficulty The difficulty of the previous Bitcoin epoch.
+ * @param currentEpochDifficulty The difficulty of the current Bitcoin epoch.
+ * @dev The block headers must come from Bitcoin epochs with difficulties marked
+ *      by the previous and current difficulties. If a Bitcoin difficulty relay
+ *      is used to provide these values and the relay is up-to-date, only the
+ *      recent block headers will pass validation. Block headers older than the
+ *      current and previous Bitcoin epochs will fail.
+ * @throws {Error} If any of the block headers are invalid, or if the block
+ *         header chain is not continuous.
+ * @returns An empty return value.
+ */
+export function validateBlockHeadersChain(
+  blockHeaders: BlockHeader[],
+  previousEpochDifficulty: BigNumber,
+  currentEpochDifficulty: BigNumber
+) {
+  let requireCurrentDifficulty: boolean = false
+  let previousBlockHeaderHash: Hex = Hex.from("00")
+
+  for (let index = 0; index < blockHeaders.length; index++) {
+    const currentHeader = blockHeaders[index]
+
+    // Check if the current block header stores the hash of the previously
+    // processed block header. Skip the check for the first header.
+    if (index !== 0) {
+      if (
+        !previousBlockHeaderHash.equals(currentHeader.previousBlockHeaderHash)
+      ) {
+        throw new Error("Invalid headers chain")
+      }
+    }
+
+    const difficultyTarget = bitsToTarget(currentHeader.bits)
+
+    const currentBlockHeaderHash = computeHash256(
+      serializeBlockHeader(currentHeader)
+    )
+
+    // Ensure the header has sufficient work.
+    if (hashLEToBigNumber(currentBlockHeaderHash).gt(difficultyTarget)) {
+      throw new Error("Insufficient work in the header")
+    }
+
+    // Save the current block header hash to compare it with the next block
+    // header's previous block header hash.
+    previousBlockHeaderHash = currentBlockHeaderHash
+
+    // Check if the stored block difficulty is equal to previous or current
+    // difficulties.
+    const difficulty = targetToDifficulty(difficultyTarget)
+
+    if (previousEpochDifficulty.eq(1) && currentEpochDifficulty.eq(1)) {
+      // Special case for Bitcoin Testnet. Do not check block's difficulty
+      // due to required difficulty falling to `1` for Testnet.
+      continue
+    }
+
+    if (
+      !difficulty.eq(previousEpochDifficulty) &&
+      !difficulty.eq(currentEpochDifficulty)
+    ) {
+      throw new Error(
+        "Header difficulty not at current or previous Bitcoin difficulty"
+      )
+    }
+
+    // Additionally, require the header to be at current difficulty if some
+    // headers at current difficulty have already been seen. This ensures
+    // there is at most one switch from previous to current difficulties.
+    if (requireCurrentDifficulty && !difficulty.eq(currentEpochDifficulty)) {
+      throw new Error("Header must be at current Bitcoin difficulty")
+    }
+
+    // If the header is at current difficulty, require the subsequent headers to
+    // be at current difficulty as well.
+    requireCurrentDifficulty = difficulty.eq(currentEpochDifficulty)
+  }
+}
+
+/**
+ * Splits Bitcoin block headers in the raw format into an array of BlockHeaders.
+ * @param blockHeaders - string that contains block headers in the raw format.
+ * @returns Array of BlockHeader objects.
+ */
+export function splitBlockHeadersChain(blockHeaders: string): BlockHeader[] {
+  if (blockHeaders.length % 160 !== 0) {
+    throw new Error("Incorrect length of Bitcoin headers")
+  }
+
+  const result: BlockHeader[] = []
+  for (let i = 0; i < blockHeaders.length; i += 160) {
+    result.push(
+      deserializeBlockHeader(Hex.from(blockHeaders.substring(i, i + 160)))
+    )
+  }
+
+  return result
 }
